+++ main.go
@@ -31,6 +31,7 @@ import (
 	controllers "github.com/converged-computing/metrics-operator/controllers/metric"
 
 	// Metrics are registered here! Importing registers once
+	_ "github.com/converged-computing/metrics-operator/pkg/metrics/application"
 	_ "github.com/converged-computing/metrics-operator/pkg/metrics/io"
 	_ "github.com/converged-computing/metrics-operator/pkg/metrics/network"
 	_ "github.com/converged-computing/metrics-operator/pkg/metrics/perf"